WASM support clap
